:building_construction: Improved import of the notes dump (fix #1)
Refactored the import script to delete notes that are included in
the local database but are not present in the notes dump anymore.
In addition to that it is now possible to run the import script again
without needing to clean the whole database before.
Also added another additional script to detect and delete notes that should not
be included in the database because they are not in the dump anymore.